Add Livetest step to CircleCI configuration (#144)
* Add Livetest step to CircleCI configuration.
Completely disable Python env caching. (Caching was not a good idea, because `requirements` files do not pin the version of every recursive dependency. Also, because we do not pin the sibling repositories (ce and 'extras') to specific tags/commits. Not a loss, because build speedup was insignificant.)
Switch to machine executor.

* Fail the build if livetest failed.

* Trigger build

* Exclude @tfestimator, @tfkeras from CI